requirements.txt
streamlit
python-docx
PyPDF2
sentence-transformers
chromadb
torch
